third_party/blink/web_tests/platform/win/fast/forms/select-popup/popup-menu-non-latin-update-from-element-expected.txt

PASS picker.selectElement_.options[0].label is "いろはにほへと"
PASS successfullyParsed is true

TEST COMPLETE

いろはにほへと
